Add walletHasName property to domain object in getNameInfo
A name may already be present in the wallet (in hsd) even though
there are no bids or reveals yet. We add a boolean to the object
to indicate whether walletClient.getAuctionInfo() returned an object
or an error. Note that this is a separate property from the "Bob"
watchlist, and is used to offer users the Rescan Auction function.