Apply WalletSync modal and importName when placing bid on new name
If a user places a bid on an in-progress auction without manually
triggering the "Rescan Auction" process, perform the auction rescan
automatically when they place their first bid. This will ensure that
all historical bids are loaded into the wallet so the user gets a
complete view of the entire auction.