Rescan auction for existing bids (#117)
* Upgrade hsd to include rpc importname and wdb height

* Expose rpc importname in wallet service

* Allow veritcal offset of floating tooltip message

* Add walletHasName property to domain object in getNameInfo

A name may already be present in the wallet (in hsd) even though
there are no bids or reveals yet. We add a boolean to the object
to indicate whether walletClient.getAuctionInfo() returned an object
or an error. Note that this is a separate property from the "Bob"
watchlist, and is used to offer users the Rescan Auction function.

* Rescan auction button

Will add a name to the wallet and rescan the recent chain history,
thus collecting all past bids. Button is only present if a name is
not already in the wallet (or if user has already placed a bid).

TODO:
- better UX while wallet is rescanning
- automatically perform this rescan when a user places a bid on a name
they haven't already imported.

* Expose rpc getwalletinfo in wallet service

Also fixes a missing ; after importNonce

* Introduce WalletSync modal and waitForWalletSync() action

Wallet rescans can occur when importing a seed or rescanning an
in-progress auction for existing bids. This introduces a modal
tht greys out the screen and displays a progress bar indicating the
% difference between walletDB height and blockchain height.

* Apply WalletSync modal during auction scan and seed import

* Apply WalletSync modal and importName when placing bid on new name

If a user places a bid on an in-progress auction without manually
triggering the "Rescan Auction" process, perform the auction rescan
automatically when they place their first bid. This will ensure that
all historical bids are loaded into the wallet so the user gets a
complete view of the entire auction.

* Update changelog

* Abort waitForWalletSync() if progress stalls for 5 seconds

* Async UI and error handling for importName and waitForWalletSync

The "wait for wallet sync" modal is the process we really need to
"await". The importName process will return almost immediately but
may throw an error. We can catch and display that error but we still
run that process asynchronously so the sync modal is displayed first.

* Get name actions from connect method instead of parent component

* Separate startWalletSync out from waitForWalletSync

* Refactor importname inside sendbid in ducks/names.js

* Expose stopWalletSync action for failure states

* Refactor waitForWalletSync inside sendbid in ducks/names.js

* Update app/ducks/walletActions.js

Co-authored-by: Matthew Slipper <me@matthewslipper.com>