Import modified workers module from hsd
The only real change needed to run hsd workers inside electron is
to switch the child_process method from `spawn()` to `fork()`.

The original hsd worker expects `process.argv[0]` to be the nodejs
executable itself (which, in the electron environment, it is not).
The `fork()` method explicitly spawns a nodejs process to execute
the passed-in module.