Replace confirmed & unlocked balances with unconfirmed and spendable
Unconfirmed total will include all TXs in and out of the wallet,
including unconfirmed TXs. This way the user gets instant feedback
when they send or receive. The difference between confirmed and
unconfirmed transactions is not obvious, except for the pending
transactions in the portfolio view and in the "delete unconfirmed TX"
panel.

Spendable balance is all TXs (including unconfirmed) minus value of
locked coins (incuding unconfirmed locked coins). The spendable
balance is used as the value for "Max Send"